tokens/views.py
from django.shortcuts import render
from rest_framework.views import APIView
from django.contrib.auth import authenticate
from rest_framework_simplejwt.tokens import RefreshToken
from django.http import JsonResponse
from django.shortcuts import get_object_or_404
from rest_framework.permissions import IsAuthenticated

class LoginView(APIView):
    def post(self, request):
        username = request.data.get("username")
        password = request.data.get("password")
        user = authenticate(username=username, password=password)
        if user is not None:
            refresh = RefreshToken.for_user(user)
            return JsonResponse(
                {"refresh": str(refresh), "access": str(refresh.access_token)}
            )
        else:
            return JsonResponse({"error": "Invalid Credentials"}, status=401)

class RestrictedView(APIView):
    permission_classes = [IsAuthenticated]
    def get(self,request):
        return JsonResponse({"response":"You are Valid User"})