requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
black==23.7.0
boto3==1.28.26
botocore==1.31.26
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.3
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.50.0
django-bulk-update==2.2.0
django-bulk-update-or-create==0.3.0
django-cors-headers==4.2.0
django-enum-choices==2.1.4
django-jsonfield==1.4.1
django-linear-migrations==2.11.0
django-redis==5.3.0
django-split-settings==1.2.0
django-storages==1.13.2
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.1
lxml==4.9.3
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
phonenumbers==8.13.18
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pypdf==3.15.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.0
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==5.5.0
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
twilio==7.17.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
xhtml2pdf==0.2.11
XlsxWriter==3.1.2
yarl==1.9.2
zope.interface==6.0
